Crate merkle_sigs

merkle_sigs implements Merkle signatures in Rust.

Structs

MerklePublicKey
A wrapper struct around a Lamport public key that implements Hashable.

Proof
An inclusion proof represent the fact that a value is a member of a MerkleTree with root hash root_hash, and hash function algorithm.

PublicKey
A one-time signing public key

Functions

sign_data_vec
Signs the entries of the data vector

verify_data_vec_signature
Verifies the signature of the data. Returns an error if data couldn't be verified.

Type Definitions

MerkleSignature
A type alias defining a Merkle signature. That includes both the Lamport leaf signature and inclusion proof.

MerkleSignedData
A type alias defining Merkle signed data. That includes the data being signed along with the signature.